Using long instead of int for the answers

+++ src/main/java/Day.java
@@ -30,21 +30,21 @@ public abstract class Day {
 
     private void run(String input) {
         long start1 = System.currentTimeMillis();
-        int res1 = part1(input);
+        long res1 = part1(input);
         long time1 = System.currentTimeMillis() - start1;
         System.out.printf("%s: %d%n", part1Label, res1);
         System.out.printf("Part 1 took %d ms%n", time1);
 
         long start2 = System.currentTimeMillis();
-        int res2 = part2(input);
+        long res2 = part2(input);
         long time2 = System.currentTimeMillis() - start2;
         System.out.printf("%s: %d%n", part2Label, res2);
         System.out.printf("Part 2 took %d ms%n", time2);
     }
 
-    abstract int part1(String input);
+    abstract long part1(String input);
 
-    abstract int part2(String input);
+    abstract long part2(String input);
 
     private String readFile(String fileName) {
         try {
